+++ apps/api/src/controllers/user.js
@@ -41,7 +41,7 @@ class UserController {
 
 	async add(req, res) {
 		try {
-			const user = await this.userService.add({ body: req.body, files: req.files })
+			const user = await this.userService.add({ body: req.body, files: req.files, authorizationHeader: req.headers.authorization })
 			const response = this.handleResponseHandler.buildResponse(user)
 	
 			res.status(response[this.responseBody.STATUS]).json(response)
@@ -55,7 +55,7 @@ class UserController {
 	
 	async findOne(req, res) {
 		try {
-			const user = await this.userService.findOne({ id: req.params.id, query: req.query })
+			const user = await this.userService.findOne({ id: req.params.id, query: req.query, authorizationHeader: req.headers.authorization })
 			const response = this.handleResponseHandler.buildResponse(user)
 	
 			res.status(response[this.responseBody.STATUS]).json(response)
@@ -69,7 +69,7 @@ class UserController {
 	
 	async list(req, res) {
 		try {
-			const user_list = await this.userService.list({ query: req.query })
+			const user_list = await this.userService.list({ query: req.query, authorizationHeader: req.headers.authorization })
 			const response = this.handleResponseHandler.buildResponse(user_list)
 	
 			res.status(response[this.responseBody.STATUS]).json(response)
@@ -97,7 +97,7 @@ class UserController {
 
 	async replace(req, res) {
 		try {
-			const user = await this.userService.replace({ body: req.body, id: req.params.id, files: req.files })
+			const user = await this.userService.replace({ body: req.body, id: req.params.id, files: req.files, authorizationHeader: req.headers.authorization })
 			const response = this.handleResponseHandler.buildResponse(user)
 	
 			res.status(response[this.responseBody.STATUS]).json(response)
@@ -111,7 +111,7 @@ class UserController {
 	
 	async remove(req, res) {
 		try {
-			const user = await this.userService.remove({ id: req.params.id })
+			const user = await this.userService.remove({ id: req.params.id, authorizationHeader: req.headers.authorization })
 			const response = this.handleResponseHandler.buildResponse(user)
 	
 			res.status(response[this.responseBody.STATUS]).json(response)
